Added Distance calculator app
